remove skyring part
  remove skyring directories from tests
  move all tendrl tests one level/directory up
  remove empty tendrl directory from tests
  remove skyring directories from modules
