+++ usmqe_tests/api/others/test_authentication.py
