+++ usmqe_tests/api/others/test_ping.py
