--- usmqe_tests/web/skyring/readme
@@ -1,1 +1,0 @@
-This directory is for web tests of skyring (RHSC 2.0)
